refactor: move camera engine logic to EcsSystem
